pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiomysql>=0.3.2",
    "alembic>=1.18.4",
    "fastapi[all]>=0.128.0",
    "passlib>=1.7.4",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "redis>=7.1.1",
    "sqlmodel>=0.0.31",
    "uvicorn[standard]>=0.40.0",
]

[tool.pytest.ini_options]
pythonpath = "."

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
